Boxplot using ggplot2
The script makes a simple boxplot of 2 input variables. In this script variable 1 was mapped as a function of variable 2 using the R function factor() inside the arguments of ggplot().  I also included the option of showing the data as dots on top of each barplot.
This adds a small amount of random variation to the location of each point, useful for handling overplotting.